web - Generates a debian package as artifact
This follows the issue #405, and is willing to harmonize the way the
artifacts are built in the other geOrchestra modules.

My suggestion for #405 would be to use the master branch and regularly
tag "2x.z.t-georchestra" each time we need to release a new version of
geOrchestra.

The following PR will the work this way. Using the following maven
command into the web module:

```
% mvn clean package deb:package -PdebianPackage
```

This will produce the following debian package:

```
georchestra-mapstore_<closest tag name>-<number of commits since the previous tag>_all.deb
```

depending on how the CI is configured, we might want to only generate
packages against a tag, not for each commits, as the previous packages
will be discarded by reprepro when publishing.