deb - fixing webapp generation
This fixes a previous work from PR407 ; the maven-resources-plugins
copies resources to target/mapstore, but the debianPackage maven profile
expects to generate a war file from target/mapstore-generic (as stated
in the previous mentioned PR, this is a convention used in the other
webapps from geOrchestra).

This commit suggests to externalize the target directory for resources
to a variable, which is then overriden in the maven profile.

It also adds extra resources to be excluded while copying. e.g if the
mavne-frontend-plugin is used, then a node/ subdirectory is created, we
don't need it in the generated webapp, so it has been added to the
excludes list, same for coverage.

Note: since no CI has been set up yet to generate the debian package,
this is probably the reason why this bug hasn't been tracked down.

tests: checking against the mapstore.war coming from the last
geosolutions-it/mapstore:latest docker image, there also seems to
contain an unecessary web/ directory:

```
    testing: web/                     OK
    testing: web/target/mapstore/embedded.html   OK
    testing: web/target/mapstore/index.html   OK
    testing: web/target/mapstore/     OK
    testing: web/target/              OK
```

both embedded.html & index.html are nested in the dist/ subdirectory of
the war file anyway.